Change codebase to double-quotes.

+++ .eslintrc.js
@@ -40,7 +40,7 @@ module.exports = {
     }],
 
     "quotes": "off",
-    "@typescript-eslint/quotes": ["error", "single", { "allowTemplateLiterals": true, "avoidEscape": true }],
+    "@typescript-eslint/quotes": ["error", "double", { "allowTemplateLiterals": true, "avoidEscape": true }],
     "semi": "off",
     "@typescript-eslint/semi": ["error", "always", { "omitLastInOneLineBlock": true }],
     "comma-dangle": ["error", {
